--- results/cdf_results.csv
@@ -1,11 +1,0 @@
-cdf,count
-0.059,0.17
-0.1535,0.24000000000000002
-0.28800000000000003,0.31000000000000005
-0.28800000000000003,0.38
-0.47100000000000003,0.45000000000000007
-0.6622,0.52
-0.6622,0.5900000000000001
-0.8189,0.66
-0.9325,0.7300000000000001
-1.0,0.8
