+++ src/windtexter.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.1
 Name: windtexter
-Version: 0.1.dev61+ge77aa1c.d20230112
+Version: 0.1.dev63+gdac3d7a.d20230113
 Summary: Cyber attack simulation model
 Home-page: https://github.com/Bonface-Osoro/windtexter.git
 Author: Bonface Osoro/Ed Oughton
