+++ src/windtexter.egg-info/SOURCES.txt
@@ -154,15 +154,16 @@ dist/globalsat-0.1.dev5+g3ee29de.d20221121-py3.9.egg
 dist/windtexter-0.1.dev22+gbc3d9c8.d20221205-py3.10.egg
 dist/windtexter-0.1.dev5+g3ee29de.d20221121-py3.9.egg
 dist/windtexter-0.1.dev6+g173f4c7.d20221121-py3.9.egg
+dist/windtexter-0.1.dev61+ge77aa1c.d20230112-py3.10.egg
+dist/windtexter-0.1.dev61+ge77aa1c.d20230112-py3.9.egg
 dist/windtexter-0.1.dev8+gad19934-py3.9.egg
 docs/Box_model.png
 docs/ecdfs.png
 docs/windtexter.png
 results/.DS_Store
 results/cdf_results.csv
-results/results.csv
 results/signal_results.csv
-results/sim_results.csv
+results/windtexter_results.csv
 scripts/Mesa Introduction.ipynb
 scripts/model.py
 scripts/run.py
@@ -172,7 +173,6 @@ src/.DS_Store
 src/.ipynb_checkpoints/model-checkpoint.ipynb
 src/__pycache__/agents.cpython-39.pyc
 src/windtexter/.DS_Store
-src/windtexter/agents.py
 src/windtexter/link_budget.py
 src/windtexter.egg-info/PKG-INFO
 src/windtexter.egg-info/SOURCES.txt
@@ -187,14 +187,12 @@ src/windtexter/__pycache__/link_budget.cpython-310.pyc
 vis/.DS_Store
 vis/plots.R
 vis/plots.py
-vis/results.csv
 vis/signal_results.csv
-vis/sim_results.csv
 vis/vis.py
 vis/wind_plots.R
+vis/windtexter_results.csv
 vis/figures/.DS_Store
 vis/figures/density_plots.png
 vis/figures/ecdfs.png
 vis/figures/loss_profile.png
-vis/figures/pub_capacity_profile.tiff
 vis/figures/wind_ecdfs.png
