+++ vis/wind_plots.R
@@ -2,6 +2,7 @@ library(ggpubr)
 library(ggplot2)
 library(dplyr)
 library(tidyverse)
+library(ggfortify)
 
 
 # Set default folder
@@ -17,7 +18,7 @@ data <- read.csv(file.path(folder, "windtexter_results.csv"))
 ecdf_intc <- ggplot(data,
                         aes(x = probability, 
                             color = windtexter)) +
-  stat_ecdf() +
+  ggdistribution() +
   scale_fill_brewer(palette = "Paired") +
   theme(legend.position = "right") +
   labs(
@@ -159,15 +160,3 @@ png(
 )
 print(density_plots)
 dev.off()
-
-
-
-
-
-
-
-
-
-
-
-
